Extended RDF Model with N-Quads Syntax.
Instead of the N-Triples syntax the RDF model
know uses the N-Quads syntax, which is an enhancement
of the N-Triple syntax. This allows us to use the concept
of named graphs.

A new model called Dataset is introduced, that represents a set of
graphs, including one default graph and n named graphs.

The triplify service uses the extended RDF features and allows
you to specify a set of addional named graphs to be created, when
triplifying your data.